+++ .gitignore
@@ -31,5 +31,3 @@ build/
 
 ### VS Code ###
 .vscode/
-
-*.db
