+++ gateway/src/main/java/ru/practicum/shareit/exception/EmailDuplicateException.java
@@ -1,0 +1,8 @@
+package ru.practicum.shareit.exception;
+
+public class EmailDuplicateException extends RuntimeException {
+
+    public EmailDuplicateException(String message) {
+        super(message);
+    }
+}
