+++ gateway/src/main/java/ru/practicum/shareit/exception/EntityNotFoundException.java
@@ -1,0 +1,9 @@
+package ru.practicum.shareit.exception;
+
+public class EntityNotFoundException extends RuntimeException {
+
+    public EntityNotFoundException(String message) {
+        super(message);
+    }
+}
+
