+++ gateway/src/main/java/ru/practicum/shareit/exception/ErrorResponse.java
@@ -1,0 +1,12 @@
+package ru.practicum.shareit.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class ErrorResponse {
+    private String className;
+    private String message;
+
+}
