+++ gateway/src/main/java/ru/practicum/shareit/exception/UnsupportedStatusException.java
@@ -1,0 +1,7 @@
+package ru.practicum.shareit.exception;
+
+public class UnsupportedStatusException extends RuntimeException {
+    public UnsupportedStatusException(String message) {
+        super(message);
+    }
+}
