+++ server/src/main/resources/application.properties
@@ -1,15 +1,14 @@
 server.port=9090
-
 spring.jpa.hibernate.ddl-auto=none
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQL10Dialect
 spring.jpa.properties.hibernate.format_sql=true
 spring.sql.init.mode=always
-
 #---
 spring.datasource.driverClassName=org.postgresql.Driver
-spring.datasource.url=jdbc:postgresql://localhost:5432/shareit
-spring.datasource.username=root
-spring.datasource.password=root
+spring.datasource.driverClassName=org.postgresql.Driver
+spring.datasource.url=${SPRING_DATASOURCE_URL}
+spring.datasource.username=${POSTGRES_USER}
+spring.datasource.password=${POSTGRES_PASSWORD}
 #---
 spring.config.activate.on-profile=ci,test
 spring.datasource.driverClassName=org.h2.Driver
